move MaxAggregationResult implementation into own file
