+++ searchlib/src/vespa/searchlib/aggregation/CMakeLists.txt
@@ -10,6 +10,7 @@ vespa_add_library(searchlib_aggregation OBJECT
     hit.cpp
     hitlist.cpp
     hitsaggregationresult.cpp
+    maxaggregationresult.cpp
     modifiers.cpp
     quantile_aggregation_result.cpp
     rawrank.cpp
