+++ searchlib/src/vespa/searchlib/aggregation/maxaggregationresult.h
@@ -7,21 +7,30 @@
 
 namespace search::aggregation {
 
+/**
+ * Aggregator that keeps the maximum value.
+ */
 class MaxAggregationResult : public AggregationResult {
 public:
     using SingleResultNode = expression::SingleResultNode;
-    DECLARE_AGGREGATIONRESULT(MaxAggregationResult);
-    MaxAggregationResult();
-    MaxAggregationResult(const SingleResultNode& max);
-    ~MaxAggregationResult();
-    void visitMembers(vespalib::ObjectVisitor& visitor) const override;
-    const SingleResultNode& getMax() const { return *_max; }
 
 private:
-    const ResultNode& onGetRank() const override { return getMax(); }
+    SingleResultNode::CP _max;
+
+public:
+    DECLARE_AGGREGATIONRESULT(MaxAggregationResult);
+
+    MaxAggregationResult();
+    explicit MaxAggregationResult(const SingleResultNode& max);
+    ~MaxAggregationResult() override;
+
+    void visitMembers(vespalib::ObjectVisitor& visitor) const override;
+    [[nodiscard]] const SingleResultNode& getMax() const { return *_max; }
+
+private:
+    [[nodiscard]] const ResultNode& onGetRank() const override { return getMax(); }
     void onPrepare(const ResultNode& result) override;
     void initForUnitTest(const ResultNode& result) override;
-    SingleResultNode::CP _max;
 };
 
 } // namespace search::aggregation
